Refactored configuration imports and splits
Mainly, the configuration imports in various modules have been refactored for better organization. Configurations have now been split into 'smib.slack.config' and 'smib.webserver.config' files, ensuring that the relevant configurations are loaded for each specific module. Docker build arguments have also been added to exclude unnecessary directories during the build process, making the build more efficient.

There is a core smib.common.config which holds configuration necessary to both parts of the app.

Docker compose file/build file have had an argument added to remove the other containers source, this is to stop an dependancy injection issue. Also makes the respective images a tad smaller
