+++ smib/slack/__main__.py
@@ -8,7 +8,7 @@ from slack_bolt import BoltRequest
 from slack_bolt.adapter.socket_mode import SocketModeHandler
 
 from smib.slack.logging_injector import inject_logger_to_slack_context
-from smib.common.config import SLACK_APP_TOKEN, SLACK_BOT_TOKEN, APPLICATION_NAME, ROOT_DIRECTORY
+from smib.slack.config import SLACK_APP_TOKEN, SLACK_BOT_TOKEN, APPLICATION_NAME, ROOT_DIRECTORY
 from smib.slack.websocket import server as websocket_server
 from smib.slack.error import handle_errors
 from injectable import Autowired, load_injection_container, autowired, injectable_factory, inject
