+++ smib/slack/config.py
@@ -1,0 +1,16 @@
+from smib.common.config import *
+
+SLACK_APP_TOKEN = config('SLACK_APP_TOKEN')
+SLACK_BOT_TOKEN = config('SLACK_BOT_TOKEN')
+
+MONGO_DB_HOST = config('MONGO_DB_HOST', default='localhost')
+MONGO_DB_PORT = config('MONGO_DB_PORT', default=27017, cast=int)
+
+MONGO_DB_DEFAULT_DB = config("MONGO_DB_DEFAULT_DB", default="smib_default")
+MONGO_DB_CONNECT_TIMEOUT_SECONDS = config("MONGO_DB_CONNECT_TIMEOUT_SECONDS", default=5, cast=int)
+
+MONGO_DB_URL = f"mongodb://{MONGO_DB_HOST}:{MONGO_DB_PORT}/"
+
+PLUGINS_DIRECTORY = config('PLUGINS_DIRECTORY', default=ROOT_DIRECTORY / 'slack' / 'plugins', cast=Path)
+
+SPACE_OPEN_ANNOUNCE_CHANNEL_ID = config('SPACE_OPEN_ANNOUNCE_CHANNEL_ID', default='C06UDPLQRP1')
